# docs/swagger.yaml
swagger: '2.0'
info:
  title: Gryphon API
  description: Store and load case management fragments using BPMN and a JSON API.
  version: 1.0.0
schemes:
  - http
basePath: /api
produces:
  - application/json
paths:
...
  /fragment/{id}:
    get:
      summary: Returns an fragment.
      description: |
        This endpoint returns a fragment based on the given ID.
      parameters:
        - name: id
          in: path
          description: The fragments id.
          required: true
          type: number
          format: integer
        - name: deliver_xml
          in: query
          description: If false the the endpoint will not deliver the fragments xml.
          required: false
          type: boolean
          format: boolean
      tags:
        - Fragment
      responses:
        '200':
          description: The fragment if found.
        '404':
          description: Fragment not found.
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: Updates a fragment.
      description: |
        This endpoint updates a fragment based on the given ID. The repository
        doesn't care wether the ID in the path and the ID in the fragment match,
        it always uses the fragment. The fragments revision will be increased if
        it differs (XML based) from earlier versions. All scenarios that are
        associated with the given fragment will also get upated
        (including a new revision). The repository will also parse the fragment
        into an Object, by parsing the given XML. The parsed structure will be
        used for validation and the structure endpoint.
      parameters:
        - name: id
          in: path
          description: The fragments id.
          required: true
          type: number
          format: integer
        - name: fragment
          in: body
          description: The fragment.
          required: true
          schema:
            $ref: '#/definitions/Fragment'
      tags:
        - Fragment
      responses:
        '200':
          description: The fragment if found and update was succesfull.
          schema:
            $ref: '#/definitions/Fragment'
        '404':
          description: Fragment not found.
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      summary: Deletes a fragment.
      description: |
        This endpoint deletes the fragment with this id.
      parameters:
        - name: id
          in: path
          description: The fragments id.
          required: true
          type: number
          format: integer
      tags:
        - Fragment
      responses:
        '200':
          description: The fragment if found.
          schema:
            $ref: '#/definitions/Fragment'
        '404':
          description: Fragment not found.
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /fragment/{id}/structure:
    get:
      summary: Returns a fragments BPMN-structure
      description: |
        This returns the BPMN structure of a given fragment. This is done by
        parsing the XMl stripping all attributes that are not used for the
        data structure (attributes that are just used for visualisation).
      parameters:
        - name: id
          in: path
          description: The fragments id.
          required: true
          type: number
          format: integer
      tags:
        - Fragment
      responses:
        '200':
          description: The fragments structure if found.
          schema:
            $ref: '#/definitions/FragmentStructureObject'
        '404':
          description: Fragment not found.
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /fragment/{id}/validate:
    get:
      summary: Validates a fragment
      description: |
        This validates a fragment and returns all messages generated by the validator.
        A fragment is validated according to the following rules:
        1. The fragment is structural sound.
        2. The fragment doesn't reference any invalid dataclasses or attributes.
        3. There are no state-transitions that are not allowed according to the
        dataclass-id
        4. There are no events without Event-Query
        5. There are no duplicate events after event-based-gateways.
        6. There are no events that are not support in the execution engine.
      parameters:
        - name: id
          in: path
          description: The fragments id.
          required: true
          type: number
          format: integer
      tags:
        - Fragment
      responses:
        '200':
          description: The list of messages.
          schema:
            $ref: '#/definitions/MessageList'
        '404':
          description: Fragment not found.
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /fragment/{id}/xml:
    get:
      summary: Returns a fragments BPMN-xml
      description: |
        This returns the BPMN-xml of a given fragment.
      parameters:
        - name: id
          in: path
          description: The fragments id.
          required: true
          type: number
          format: integer
      tags:
        - Fragment
      responses:
        '200':
          description: The fragment-xml if found.
        '404':
          description: Fragment not found.
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  /fragment:
    get:
      summary: Returns a list of all available fragments.
      description: |
        Returns a list of all fragments. This list might be filtered by searching
        using a name.
      parameters:
        - name: query
          in: query
          description: Might be used to filter the fragments by name.
          required: false
          type: string
          format: string
      tags:
      - Fragment
      responses:
        '200':
          description: The list of fragments.
          schema:
            $ref: '#/definitions/FragmentList'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: Creates a new scenario
      description: |
        This endpoint stores a new fragment on the server. If the body contains
        contains no content, the default fragment xml is used.
      parameters:
        - name: scenario
          in: body
          description: The fragment object.
          required: true
          schema:
            $ref: '#/definitions/Fragment'
      tags:
        - Fragment
      responses:
        '200':
          description: The updated scenario, maybe with updated revision.
          schema:
            $ref: '#/definitions/Fragment'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
...
  /fragment/{id}/assocdomainmodel:
    get:
      summary: Returns the domainmodel associated with this fragment.
      description: |
        Returns the domainmodel that belongs to the scenario this fragment belongs to.
      parameters:
        - name: id
          in: path
          description: The domain models id.
          type: number
          format: integer
          required: true
      tags:
        - DomainModel
        - Fragment
      responses:
        '200':
          description: The domain model if found.
          schema:
            $ref: '#/definitions/DomainModel'
        '404':
          description: Model not found.
          schema:
            $ref: '#/definitions/Error'
        '500':
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
definitions:
...
  Fragment:
    type: object
    properties:
      name:
        type: string
        description: 'The fragments name.'
      _id:
        type: number
        format: integer
        description: 'The fragments unique identifier.'
      content:
        type: string
        format: integer
        description: 'The BPMN-XML used to describe the content.'
      revision:
        type: number
        format: integer
        description: 'The fragments revision. Gets increased everytime the fragment is updated.'
  FragmentStructureObject:
    type: object
  DomainModel:
    type: object
    required:
      - name
      - dataclasses
    properties:
      name:
        type: string
        description: 'The name of the domain model.'
      _id:
        type: number
        format: long
        description: 'The unique identifier of the domain model.'
      revision:
        type: number
        format: integer
        description: 'The domain models revision. Gets increased everytime the domain model is updated.'
      dataclasses:
        type: array
        description: 'The data classes belonging to the domain model.'
        items:
          $ref: '#/definitions/DataClass'
  DataClass:
    type: object
    properties:
      name:
        type: string
        description: 'The name of the data class.'
      _id:
        type: number
        format: long
        description: 'The unique identifier the data class.'
      is_root:
        type: boolean
        description: 'Describes whether the data class is the root element of the domain model.'
      is_event:
        type: boolean
        description: 'Describes whether the data class represents event data.'
      attributes:
        type: array
        description: 'The attributes of the data class.'
        items:
          $ref: '#/definitions/DataClassAttribute'
  DataClassAttribute:
    type: object
    properties:
      name:
        type: string
        description: 'The name of the attribute.'
      datatype:
        type: string
        description: 'The datatype of the attribute.'
  MessageList:
    type: array
    description: 'A list of messages.'
    items:
      $ref: '#/definitions/Message'
  Message:
    type: object
    properties:
      text:
        type: string
        description: 'The message for this error. This should be displayed to the user.'
      type:
        type: string
        description: 'The type of the error.'
...
  FragmentList:
    type: array
    description: 'All fragments found.'
    items:
      $ref: '#/definitions/Fragment'
...
  Error:
    type: object
...